refactor(extension): add code improvements
