+++ apps/browser-extension-wallet/src/lib/scripts/background/services/userIdService.ts
@@ -11,14 +11,14 @@ import {
 import { USER_ID_SERVICE_BASE_CHANNEL, UserIdService as UserIdServiceInterface } from '@lib/scripts/types';
 import randomBytes from 'randombytes';
 import { userIdServiceProperties } from '../config';
-import { getChainNameByNetworkMagic } from '@src/utils/chain';
+import { getChainNameByNetworkMagic } from '@src/utils/get-chain-name-by-network-magic';
 
 // eslint-disable-next-line no-magic-numbers
 export const SESSION_LENGTH = Number(process.env.SESSION_LENGTH_IN_SECONDS || 1800) * 1000;
 export const USER_ID_BYTE_SIZE = 8;
 
 export class UserIdService implements UserIdServiceInterface {
-  private userId?: string;
+  private randomizedUserId?: string;
   private walletBasedUserId?: string;
   private sessionTimeout?: NodeJS.Timeout;
   private userIdRestored = false;
@@ -30,17 +30,16 @@ export class UserIdService implements UserIdServiceInterface {
     private sessionLength: number = SESSION_LENGTH
   ) {}
 
-  async getwalletBasedUserId(networkMagic: Wallet.Cardano.NetworkMagic): Promise<string | undefined> {
+  private async getWalletBasedUserId(networkMagic: Wallet.Cardano.NetworkMagic): Promise<string | undefined> {
     const { keyAgentsByChain, usePersistentUserId } = await this.getStorage();
 
     if (!keyAgentsByChain) {
       console.debug('[ANALYTICS] Key agents not found - Wallet not created yet');
-      return this.walletBasedUserId;
+      return undefined;
     }
 
     if (!usePersistentUserId) {
-      this.walletBasedUserId = undefined;
-      return this.walletBasedUserId;
+      return undefined;
     }
 
     if (!this.walletBasedUserId) {
@@ -50,6 +49,7 @@ export class UserIdService implements UserIdServiceInterface {
     }
 
     console.debug(`[ANALYTICS] getwalletBasedUserId() called (current Wallet Based ID: ${this.walletBasedUserId})`);
+    // eslint-disable-next-line consistent-return
     return this.walletBasedUserId;
   }
 
@@ -59,18 +59,17 @@ export class UserIdService implements UserIdServiceInterface {
       await this.restoreUserId();
     }
 
-    if (!this.userId) {
+    if (!this.randomizedUserId) {
       console.debug('[ANALYTICS] User ID not found - generating new one');
-      this.userId = randomBytes(USER_ID_BYTE_SIZE).toString('hex');
+      this.randomizedUserId = randomBytes(USER_ID_BYTE_SIZE).toString('hex');
     }
 
-    console.debug(`[ANALYTICS] getId() called (current ID: ${this.userId})`);
-
-    return this.userId;
+    console.debug(`[ANALYTICS] getId() called (current ID: ${this.randomizedUserId})`);
+    return this.randomizedUserId;
   }
 
   async getUserId(networkMagic: Wallet.Cardano.NetworkMagic): Promise<string> {
-    const walletBasedId = await this.getwalletBasedUserId(networkMagic);
+    const walletBasedId = await this.getWalletBasedUserId(networkMagic);
 
     if (!walletBasedId) {
       return await this.getRandomizedUserId();
@@ -80,14 +79,14 @@ export class UserIdService implements UserIdServiceInterface {
   }
 
   async getAliasProperties(networkMagic: Wallet.Cardano.NetworkMagic): Promise<{ alias: string; id: string }> {
-    const id = await this.getwalletBasedUserId(networkMagic);
+    const id = await this.getWalletBasedUserId(networkMagic);
     const alias = await this.getRandomizedUserId();
     return { alias, id };
   }
 
   async clearId(): Promise<void> {
     console.debug('[ANALYTICS] clearId() called');
-    this.userId = undefined;
+    this.randomizedUserId = undefined;
     this.walletBasedUserId = undefined;
     this.clearSessionTimeout();
     await this.clearStorage(['userId', 'usePersistentUserId']);
@@ -120,7 +119,7 @@ export class UserIdService implements UserIdServiceInterface {
 
     if (usePersistentUserId) {
       console.debug('[ANALYTICS] Restoring user ID from extension storage');
-      this.userId = userId;
+      this.randomizedUserId = userId;
     }
 
     this.userIdRestored = true;
@@ -131,7 +130,7 @@ export class UserIdService implements UserIdServiceInterface {
       return;
     }
     this.sessionTimeout = setTimeout(() => {
-      this.userId = undefined;
+      this.randomizedUserId = undefined;
       console.debug('[ANALYTICS] Session timed out');
     }, this.sessionLength);
   }
@@ -143,6 +142,7 @@ export class UserIdService implements UserIdServiceInterface {
 
   private generateWalletBasedUserId(extendedAccountPublicKey: Wallet.Crypto.Bip32PublicKeyHex) {
     console.debug('[ANALYTICS] Wallet based ID not found - generating new one');
+    // by requirement, we want to hash the extended account public key twice
     const hash = hashExtendedAccountPublicKey(extendedAccountPublicKey);
     return hashExtendedAccountPublicKey(hash);
   }
