+++ apps/browser-extension-wallet/src/utils/get-chain-name-by-network-magic.ts
@@ -1,0 +1,4 @@
+import { Wallet } from '@lace/cardano';
+
+export const getChainNameByNetworkMagic = (networkMagic: Wallet.Cardano.NetworkMagics): Wallet.ChainName =>
+  Object.entries(Wallet.Cardano.ChainIds).find(([, ids]) => ids.networkMagic === networkMagic)[0] as Wallet.ChainName;
