+++ apps/browser-extension-wallet/src/utils/mocks/test-helpers.tsx
@@ -621,8 +621,7 @@ export const userIdServiceMock: Record<keyof UserIdService, jest.Mock> = {
   clearId: jest.fn(),
   getRandomizedUserId: jest.fn(),
   getUserId: jest.fn(),
-  getAliasProperties: jest.fn(),
-  getwalletBasedUserId: jest.fn()
+  getAliasProperties: jest.fn()
 };
 
 export const matomoClientMocks: Record<keyof typeof MatomoClient.prototype, jest.Mock> = {
